# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Party {
  code: string;
  name: string;
  color: string;
}

export type ManifestationItemKind = typeof ManifestationItemKind[keyof typeof ManifestationItemKind];


export const ManifestationItemKind = {
  apoio: 'apoio',
  critica: 'critica',
} as const;

export interface ManifestationItem {
  id: string;
  label: string;
  kind: ManifestationItemKind;
  weight: number;
  tier: number;
  emoji: string;
  hint: string;
  priceCents: number;
}

export type PoliticianBaseByItem = {[key: string]: number};

export interface Politician {
  id: string;
  name: string;
  initials: string;
  role: string;
  level: string;
  region: string;
  countryCode: string;
  countryName: string;
  stateCode: string;
  stateName: string;
  /** @nullable */
  city: string | null;
  party: Party;
  score: number;
  support: number;
  criticism: number;
  trend: number;
  position: number;
  baseByItem: PoliticianBaseByItem;
}

export type ActivityItemKind = typeof ActivityItemKind[keyof typeof ActivityItemKind];


export const ActivityItemKind = {
  apoio: 'apoio',
  critica: 'critica',
} as const;

export interface ActivityItem {
  id: string;
  politicianId: string;
  politicianName: string;
  itemLabel: string;
  kind: ActivityItemKind;
  createdAt: string;
  region: string;
}

export type PoliticianProfile = Politician & {
  bio: string;
  mandates: string[];
  recentActivity: ActivityItem[];
  items: ManifestationItem[];
};

export interface ManifestationInput {
  itemId: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity?: number;
  /**
     * @maxLength 280
     * @nullable
     */
  note?: string | null;
}

export interface CheckoutLineItem {
  itemId: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity: number;
}

export interface CheckoutIntentInput {
  politicianId: string;
  /**
     * @minItems 1
     * @maxItems 20
     */
  items: CheckoutLineItem[];
}

export type CheckoutIntentStatus = typeof CheckoutIntentStatus[keyof typeof CheckoutIntentStatus];


export const CheckoutIntentStatus = {
  pending: 'pending',
  paid: 'paid',
  expired: 'expired',
  canceled: 'canceled',
} as const;

export interface CheckoutIntent {
  id: string;
  checkoutToken: string;
  politicianId: string;
  amountCents: number;
  status: CheckoutIntentStatus;
  expiresAt: string;
}

export interface Manifestation {
  id: string;
  politicianId: string;
  item: ManifestationItem;
  quantity: number;
  createdAt: string;
  /** @nullable */
  note: string | null;
}

export interface OverviewStats {
  totalPoliticians: number;
  totalManifestations: number;
  activeRegions: number;
  supportPercentage: number;
}

export type ListPoliticiansParams = {
q?: string;
level?: ListPoliticiansLevel;
region?: string;
country?: string;
state?: string;
city?: string;
sort?: ListPoliticiansSort;
};

export type ListPoliticiansLevel = typeof ListPoliticiansLevel[keyof typeof ListPoliticiansLevel];


export const ListPoliticiansLevel = {
  federal: 'federal',
  estadual: 'estadual',
  municipal: 'municipal',
} as const;

export type ListPoliticiansSort = typeof ListPoliticiansSort[keyof typeof ListPoliticiansSort];


export const ListPoliticiansSort = {
  relevancia: 'relevancia',
  apoio: 'apoio',
  critica: 'critica',
} as const;