# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: politicians
    description: Political profiles and discovery
  - name: activity
    description: Public civic activity
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /politicians:
    get:
      operationId: listPoliticians
      tags: [politicians]
      summary: List politicians
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: level
          in: query
          schema: { type: string, enum: [federal, estadual, municipal] }
        - name: region
          in: query
          schema: { type: string }
        - name: country
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [relevancia, apoio, critica] }
      responses:
        "200":
          description: Politician list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Politician" }
  /politicians/{id}:
    get:
      operationId: getPolitician
      tags: [politicians]
      summary: Get a politician profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Politician profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PoliticianProfile" }
        "404":
          description: Not found
  /politicians/{id}/manifestations:
    post:
      operationId: createManifestation
      tags: [politicians]
      summary: Register a civic manifestation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ManifestationInput" }
      responses:
        "201":
          description: Created manifestation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Manifestation" }
  /checkout-intents:
    post:
      operationId: createCheckoutIntent
      tags: [checkout]
      summary: Create an anonymous server-owned checkout intent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckoutIntentInput" }
      responses:
        "201":
          description: Checkout intent created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckoutIntent" }
        "400":
          description: Invalid checkout
  /activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: List recent public activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityItem" }
  /stats/overview:
    get:
      operationId: getOverviewStats
      tags: [activity]
      summary: Get platform overview statistics
      responses:
        "200":
          description: Overview statistics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OverviewStats" }
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Party:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        color: { type: string }
      required: [code, name, color]
    ManifestationItem:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        kind: { type: string, enum: [apoio, critica] }
        weight: { type: number }
        tier: { type: number }
        emoji: { type: string }
        hint: { type: string }
        priceCents: { type: number }
      required: [id, label, kind, weight, tier, emoji, hint, priceCents]
    Politician:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        initials: { type: string }
        role: { type: string }
        level: { type: string }
        region: { type: string }
        countryCode: { type: string }
        countryName: { type: string }
        stateCode: { type: string }
        stateName: { type: string }
        city: { type: ["string", "null"] }
        party: { $ref: "#/components/schemas/Party" }
        score: { type: number }
        support: { type: number }
        criticism: { type: number }
        trend: { type: number }
        position: { type: number }
        baseByItem:
          type: object
          additionalProperties: { type: number }
      required: [id, name, initials, role, level, region, countryCode, countryName, stateCode, stateName, city, party, score, support, criticism, trend, position, baseByItem]
    PoliticianProfile:
      allOf:
        - $ref: "#/components/schemas/Politician"
        - type: object
          properties:
            bio: { type: string }
            mandates: { type: array, items: { type: string } }
            recentActivity:
              type: array
              items: { $ref: "#/components/schemas/ActivityItem" }
            items:
              type: array
              items: { $ref: "#/components/schemas/ManifestationItem" }
          required: [bio, mandates, recentActivity, items]
    ManifestationInput:
      type: object
      properties:
        itemId: { type: string }
        quantity: { type: number, minimum: 1, maximum: 99, default: 1 }
        note: { type: ["string", "null"], maxLength: 280 }
      required: [itemId]
    CheckoutLineItem:
      type: object
      properties:
        itemId: { type: string }
        quantity: { type: number, minimum: 1, maximum: 99 }
      required: [itemId, quantity]
    CheckoutIntentInput:
      type: object
      properties:
        politicianId: { type: string }
        items:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: "#/components/schemas/CheckoutLineItem" }
      required: [politicianId, items]
    CheckoutIntent:
      type: object
      properties:
        id: { type: string }
        checkoutToken: { type: string }
        politicianId: { type: string }
        amountCents: { type: number }
        status: { type: string, enum: [pending, paid, expired, canceled] }
        expiresAt: { type: string, format: date-time }
      required: [id, checkoutToken, politicianId, amountCents, status, expiresAt]
    Manifestation:
      type: object
      properties:
        id: { type: string }
        politicianId: { type: string }
        item: { $ref: "#/components/schemas/ManifestationItem" }
        quantity: { type: number }
        createdAt: { type: string, format: date-time }
        note: { type: ["string", "null"] }
      required: [id, politicianId, item, quantity, createdAt, note]
    ActivityItem:
      type: object
      properties:
        id: { type: string }
        politicianId: { type: string }
        politicianName: { type: string }
        itemLabel: { type: string }
        kind: { type: string, enum: [apoio, critica] }
        createdAt: { type: string, format: date-time }
        region: { type: string }
      required: [id, politicianId, politicianName, itemLabel, kind, createdAt, region]
    OverviewStats:
      type: object
      properties:
        totalPoliticians: { type: number }
        totalManifestations: { type: number }
        activeRegions: { type: number }
        supportPercentage: { type: number }
      required: [totalPoliticians, totalManifestations, activeRegions, supportPercentage]