# lib/api-zod/src/generated/types/checkoutIntent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutIntentStatus } from './checkoutIntentStatus';

export interface CheckoutIntent {
  id: string;
  checkoutToken: string;
  politicianId: string;
  amountCents: number;
  status: CheckoutIntentStatus;
  expiresAt: Date;
}