# lib/api-zod/src/generated/types/checkoutIntentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutLineItem } from './checkoutLineItem';

export interface CheckoutIntentInput {
  politicianId: string;
  /**
     * @minItems 1
     * @maxItems 20
     */
  items: CheckoutLineItem[];
}