# lib/api-zod/src/generated/types/checkoutLineItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutLineItem {
  itemId: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity: number;
}